pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "bcrypt==5.0.0",
    "email-validator==2.1.0",
    "fastapi==0.104.1",
    "httpx==0.27.2",
    "passlib==1.7.4",
    "pillow>=12.2.0",
    "pydantic==2.5.0",
    "pydantic-settings==2.1.0",
    "python-dotenv==1.0.0",
    "python-jose==3.3.0",
    "python-multipart==0.0.6",
    "scikit-learn>=1.8.0",
    "supabase==2.16.0",
    "uvicorn==0.24.0",
    "xgboost>=3.2.0",
]
